requirements.txt
# ── Core API ────────────────────────────────────────────────────────────
fastapi==0.115.2
uvicorn==0.32.0
python-dotenv==1.1.0
python-multipart==0.0.12

# ── LangChain Core ─────────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0

# ── LLM Providers (install the ones you need) ──────────────────────────
langchain-anthropic>=0.3.1       # Anthropic Claude
langchain-openai>=0.2.0          # OpenAI / Azure OpenAI / OpenAI-compatible
langchain-huggingface>=0.1.0     # HuggingFace Hub & Endpoints
langchain-ollama>=0.2.0          # Ollama (local models)

# ── Browser automation ─────────────────────────────────────────────────
playwright==1.48.0

# ── Task queue ─────────────────────────────────────────────────────────
celery==5.4.0
redis==5.1.1

# ── Resume processing (OCR + PDF) ─────────────────────────────────────
pypdf2==3.0.1
paddleocr>=2.7.0

# ── Database ───────────────────────────────────────────────────────────
supabase==2.9.1

# ── Job scraper (scripts/) ─────────────────────────────────────────────
requests>=2.31.0
